mcp-recommender-system-agent/app/requirements.txt
mcp==1.9.3
python-dotenv==1.1.0
pandas==2.3.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
requests==2.32.3
sentence-transformers==4.1.0
torch==2.7.1
scikit-learn==1.7.0
transformers==4.52.4
numpy==2.3.0
scipy==1.15.3
networkx==3.5
pillow==11.2.1
datasets==3.6.0
accelerate==1.7.0
